+++ app/src/main/java/com/team/studing/UI/Notice/DialogNoticeDelete.kt
@@ -1,4 +1,4 @@
-package com.team.studing.UI.Home
+package com.team.studing.UI.Notice
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
