+++ app/src/main/res/layout/fragment_notice_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home.NoticeDetailFragment">
+    tools:context=".UI.Notice.NoticeDetailFragment">
 
     <include
         android:id="@+id/toolbar"
