+++ app/src/main/res/layout/fragment_notice_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/color_gradient_linear_10"
-    tools:context=".UI.Home.NoticeListFragment">
+    tools:context=".UI.Notice.NoticeListFragment">
 
     <include
         android:id="@+id/toolbar"
