+++ app/src/main/res/layout/fragment_scrap_notice_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/color_gradient_linear_10"
-    tools:context=".Home.ScrapNoticeListFragment">
+    tools:context=".UI.Notice.ScrapNoticeListFragment">
 
     <include
         android:id="@+id/toolbar"
