+++ app/src/main/res/layout/fragment_unread_notice.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home.UnreadNoticeFragment">
+    tools:context=".UI.Notice.UnreadNoticeFragment">
 
     <include
         android:id="@+id/toolbar"
