+++ app/src/main/res/layout/fragment_unread_notice_finish.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home.UnreadNoticeFinishFragment">
+    tools:context=".UI.Notice.UnreadNoticeFinishFragment">
 
     <include
         android:id="@+id/toolbar"
